+++ okapi/src/lib.rs
@@ -2,5 +2,7 @@
 #![deny(clippy::all)]
 
 pub type Map<K, V> = schemars::Map<K, V>;
+pub type MapEntry<'a, K, V> = schemars::MapEntry<'a, K, V>;
 
+pub mod merge;
 pub mod openapi3;